🚚 api: move usecase to session-management context
