+++ api/src/certification/session-management/domain/usecases/authorize-certification-candidate-to-resume.js
