+++ api/tests/certification/session-management/unit/domain/usecases/authorize-certification-candidate-to-resume_test.js
@@ -1,5 +1,5 @@
-import { authorizeCertificationCandidateToResume } from '../../../../lib/domain/usecases/authorize-certification-candidate-to-resume.js';
-import { domainBuilder, expect, sinon } from '../../../test-helper.js';
+import { authorizeCertificationCandidateToResume } from '../../../../../../src/certification/session-management/domain/usecases/authorize-certification-candidate-to-resume.js';
+import { domainBuilder, expect, sinon } from '../../../../../test-helper.js';
 
 describe('Unit | Domain | Use Cases | authorize-certification-candidate-to-resume', function () {
   let certificationCandidateForSupervisingRepository;
